+++ R/content.R
@@ -207,16 +207,23 @@ method(format, ContentToolRequest) <- function(
 ) {
   show <- arg_match(show)
 
-  if (length(x@arguments) == 0) {
-    call <- call2(x@name)
-  } else {
-    call <- call2(x@name, !!!x@arguments)
+  arguments <- tool_request_args(x)
+  if (is_tool_result(arguments)) {
+    # Failed to convert the arguments so just use unconverted
+    arguments <- x@arguments
   }
-  if (show == "call") {
-    return(format(call))
+  call <- call2(x@name, !!!arguments)
+  call_str <- deparse(call)
+  if (length(call_str) > 1) {
+    open <- 
+    call_str <- paste0(call_str[1], "...)")
   }
 
-  cli::format_inline("[{.strong tool request} ({x@id})]: {format(call)}")
+  if (show == "call") {
+    return(call_str)
+  }
+
+  cli::format_inline("[{.strong tool request} ({x@id})]: {call_str}")
 }
 
 #' @rdname Content