+++ tests/testthat/test-content.R
@@ -62,6 +62,35 @@ test_that("thinking has useful representations", {
   expect_snapshot(cat(contents_html(ct)))
 })
 
+test_that("ContentToolRequest shows converted arguments", {
+  my_tool <- tool(
+    function(x, y, z) {},
+    name = "my_tool",
+    description = "A tool",
+    arguments = list(
+      x = type_array(type_number()),
+      y = type_array(type_string()),
+      z = type_enum(c("a", "b", "c"))
+    )
+  )
+  content <- ContentToolRequest(
+    "id",
+    name = "my_tool",
+    arguments = list(x = c(1, 2), y = c("a", "b"), z = "a"),
+    tool = my_tool
+  )
+  expect_snapshot(cat(format(content)))
+
+  # and very long arguments are truncated
+  content <- ContentToolRequest(
+    "id",
+    name = "my_tool",
+    arguments = list(x = rep(123, 1000), y = c("a", "b"), z = "a"),
+    tool = my_tool
+  )
+  expect_snapshot(cat(format(content)))
+})
+
 test_that("ContentToolResult@error requires a string or an error condition", {
   expect_snapshot(error = TRUE, {
     ContentToolResult("id", error = TRUE)